+++ recipes/clean-up-package-cache.rb
@@ -6,3 +6,9 @@ execute 'clean package cache' do
   retries 5
   retry_delay 5
 end
+
+execute 'clean pip cache' do
+  command "/usr/bin/scl enable rh-python38 -- python -m pip cache purge"
+  retries 5
+  retry_delay 5
+end